Added Pokedex owned registration
Any Pokemon that exists in the PC when saving will be added to the
Pokedex as Owned.
Prevent loading any save data that has not yet received the Pokedex.
Enable the National Dex when saving for Ruby and Sapphire only.
The first caught form of Spinda and Unown is saved to the Pokedex.
Fixed the Pokedex numbers for Aron, Lairon, and Aggron.